Make TextMod a separate component
